Rename of group class files
