+++ geoh5py/groups/container.py
@@ -19,25 +19,16 @@ from __future__ import annotations
 
 import uuid
 
-from .group import Group, GroupType
+from .base import Group
 
 
-class GiftoolsGroup(Group):
-    """The type for a GIFtools group."""
+class ContainerGroup(Group):
+    """The type for the basic Container group."""
 
     __TYPE_UID = uuid.UUID(
-        fields=(0x585B3218, 0xC24B, 0x41FE, 0xAD, 0x1F, 0x24D5E6E8348A)
+        fields=(0x61FBB4E8, 0xA480, 0x11E3, 0x8D, 0x5A, 0x2776BDF4F982)
     )
-
-    _name = "GIFtools Project"
-    _description = "GIFtools Project"
-
-    def __init__(self, group_type: GroupType, **kwargs):
-        assert group_type is not None
-        super().__init__(group_type, **kwargs)
-
-        if self.entity_type.name == "Entity":
-            self.entity_type.name = "GIFtools Group"
+    _default_name = "Container Group"
 
     @classmethod
     def default_type_uid(cls) -> uuid.UUID:
