+++ geoh5py/groups/custom.py
@@ -19,26 +19,18 @@ from __future__ import annotations
 
 import uuid
 
-from .group import Group, GroupType
+from .base import Group
 
 
-class NoTypeGroup(Group):
-    """A group with no type."""
+class CustomGroup(Group):
+    """A custom group, for an unlisted Group type."""
 
-    __TYPE_UID = uuid.UUID(
-        fields=(0xDD99B610, 0xBE92, 0x48C0, 0x87, 0x3C, 0x5B5946EA2840)
-    )
-
-    _name = "NoType"
-    _description = "<Unknown>"
-
-    def __init__(self, group_type: GroupType, **kwargs):
-        assert group_type is not None
-        super().__init__(group_type, **kwargs)
-
-        if self.entity_type.name == "Entity":
-            self.entity_type.name = "NoType Group"
+    _default_name = "Custom Group"
 
     @classmethod
-    def default_type_uid(cls) -> uuid.UUID:
-        return cls.__TYPE_UID
+    def default_type_uid(cls) -> uuid.UUID | None:
+        """
+        Mock the default type uid for the custom group.
+        It returns a new UUID every time this class is called.
+        """
+        return uuid.uuid4()
