+++ geoh5py/groups/drillhole.py
@@ -19,7 +19,7 @@ from __future__ import annotations
 
 import uuid
 
-from .group import Group, GroupType
+from .base import Group
 
 
 class DrillholeGroup(Group):
@@ -28,13 +28,7 @@ class DrillholeGroup(Group):
     __TYPE_UID = uuid.UUID(
         fields=(0x825424FB, 0xC2C6, 0x4FEA, 0x9F, 0x2B, 0x6CD00023D393)
     )
-
-    def __init__(self, group_type: GroupType, name="Drillholes Group", **kwargs):
-        assert group_type is not None
-        super().__init__(group_type, name=name, **kwargs)
-
-        if self.entity_type.name == "Entity":
-            self.entity_type.name = name
+    _default_name = "Drillhole Group"
 
     @classmethod
     def default_type_uid(cls) -> uuid.UUID:
@@ -45,9 +39,7 @@ class IntegratorDrillholeGroup(DrillholeGroup):
     """The type for the group containing drillholes."""
 
     __TYPE_UID = uuid.UUID("{952829b6-76a2-4d0b-b908-7f8d2482dc0d}")
-
-    def __init__(self, group_type: GroupType, **kwargs):
-        super().__init__(group_type, **kwargs)
+    _default_name = "Integrator Drillhole Group"
 
     @classmethod
     def default_type_uid(cls) -> uuid.UUID:
