+++ geoh5py/groups/group_type.py
@@ -23,7 +23,7 @@ from ..shared import EntityType
 
 if TYPE_CHECKING:
     from ..workspace import Workspace
-    from . import group  # noqa: F401
+    from . import base  # noqa: F401
 
 
 class GroupType(EntityType):
