+++ geoh5py/groups/integrator.py
@@ -19,20 +19,14 @@ from __future__ import annotations
 
 import uuid
 
-from .group import Group, GroupType
+from .base import Group
 
 
 class AirborneTheme(Group):
     """The type for a INTEGRATOR Airborne Theme."""
 
     __TYPE_UID = uuid.UUID("{3d0e8578-7764-48cf-8db8-6c83d6411762}")
-
-    _name = "Airborne Theme"
-    _description = "Airborne Theme"
-
-    def __init__(self, group_type: GroupType, **kwargs):
-        assert group_type is not None
-        super().__init__(group_type, **kwargs)
+    _default_name = "Airborne Theme"
 
     @classmethod
     def default_type_uid(cls) -> uuid.UUID:
@@ -43,13 +37,7 @@ class EarthModelsTheme(Group):
     """The type for a INTEGRATOR Earth Models Theme."""
 
     __TYPE_UID = uuid.UUID("{adee3b2a-3829-11e4-a70e-fcddabfddab1}")
-
-    _name = "Earth Models Theme"
-    _description = "Earth Models Theme"
-
-    def __init__(self, group_type: GroupType, **kwargs):
-        assert group_type is not None
-        super().__init__(group_type, **kwargs)
+    _default_name = "Earth Models Theme"
 
     @classmethod
     def default_type_uid(cls) -> uuid.UUID:
@@ -60,13 +48,7 @@ class GeochemistryMineralogyTheme(Group):
     """The type for a INTEGRATOR Geochemistry & Mineralogy Theme."""
 
     __TYPE_UID = uuid.UUID("{ed00094f-3da1-485f-8c4e-b52f6f171ea4}")
-
-    _name = "Geochemistry & Mineralogy Theme"
-    _description = "Geochemistry & Mineralogy Theme"
-
-    def __init__(self, group_type: GroupType, **kwargs):
-        assert group_type is not None
-        super().__init__(group_type, **kwargs)
+    _default_name = "Geochemistry & Mineralogy Theme"
 
     @classmethod
     def default_type_uid(cls) -> uuid.UUID:
@@ -77,13 +59,7 @@ class GeochemistryMineralogyDataSet(Group):
     """The type for a INTEGRATOR Geochemistry & Mineralogy DataSet."""
 
     __TYPE_UID = uuid.UUID("{72f29283-a4f6-4fc0-a1a8-1417ce5fcbec}")
-
-    _name = "Geochemistry & Mineralogy DataSet"
-    _description = "Geochemistry & Mineralogy DataSet"
-
-    def __init__(self, group_type: GroupType, **kwargs):
-        assert group_type is not None
-        super().__init__(group_type, **kwargs)
+    _default_name = "Geochemistry & Mineralogy DataSet"
 
     @classmethod
     def default_type_uid(cls) -> uuid.UUID:
@@ -94,13 +70,7 @@ class GeophysicsTheme(Group):
     """The type for a INTEGRATOR Geophysics Theme."""
 
     __TYPE_UID = uuid.UUID("{151778d9-6cc0-4e72-ba08-2a80a4fb967f}")
-
-    _name = "Geophysics Theme"
-    _description = "Geophysics Theme"
-
-    def __init__(self, group_type: GroupType, **kwargs):
-        assert group_type is not None
-        super().__init__(group_type, **kwargs)
+    _default_name = "Geophysics Theme"
 
     @classmethod
     def default_type_uid(cls) -> uuid.UUID:
@@ -111,13 +81,7 @@ class GroundTheme(Group):
     """The type for a INTEGRATOR Ground Theme."""
 
     __TYPE_UID = uuid.UUID("{47d6f059-b56a-46c7-8fc7-a0ded87360c3}")
-
-    _name = "Ground Theme"
-    _description = "Ground Theme"
-
-    def __init__(self, group_type: GroupType, **kwargs):
-        assert group_type is not None
-        super().__init__(group_type, **kwargs)
+    _default_name = "Ground Theme"
 
     @classmethod
     def default_type_uid(cls) -> uuid.UUID:
@@ -128,13 +92,7 @@ class IntegratorProject(Group):
     """The type for a INTEGRATOR group."""
 
     __TYPE_UID = uuid.UUID("{56f6f03e-3833-11e4-a7fb-fcddabfddab1}")
-
-    _name = "Geoscience INTEGRATOR Project"
-    _description = "Geoscience INTEGRATOR Project"
-
-    def __init__(self, group_type: GroupType, **kwargs):
-        assert group_type is not None
-        super().__init__(group_type, **kwargs)
+    _default_name = "Geoscience INTEGRATOR Project"
 
     @classmethod
     def default_type_uid(cls) -> uuid.UUID:
@@ -145,13 +103,7 @@ class IntegratorGroup(Group):
     """The type for a INTEGRATOR group."""
 
     __TYPE_UID = uuid.UUID("{61449477-3833-11e4-a7fb-fcddabfddab1}")
-
-    _name = "Geoscience INTEGRATOR"
-    _description = "Geoscience INTEGRATOR"
-
-    def __init__(self, group_type: GroupType, **kwargs):
-        assert group_type is not None
-        super().__init__(group_type, **kwargs)
+    _default_name = "Geoscience INTEGRATOR"
 
     @classmethod
     def default_type_uid(cls) -> uuid.UUID:
@@ -162,13 +114,7 @@ class QueryGroup(Group):
     """The type for a INTEGRATOR Query Group."""
 
     __TYPE_UID = uuid.UUID("{85756113-592a-4088-b374-f32c8fac37a2}")
-
-    _name = "Query Group"
-    _description = "Query Group"
-
-    def __init__(self, group_type: GroupType, **kwargs):
-        assert group_type is not None
-        super().__init__(group_type, **kwargs)
+    _default_name = "Query Group"
 
     @classmethod
     def default_type_uid(cls) -> uuid.UUID:
@@ -179,13 +125,7 @@ class ObservationPointsTheme(Group):
     """The type for a INTEGRATOR Observation Points Theme."""
 
     __TYPE_UID = uuid.UUID("{f65e521c-a763-427b-97bf-d0b4e5689e0d}")
-
-    _name = "Observation Points Theme"
-    _description = "Observation Points Theme"
-
-    def __init__(self, group_type: GroupType, **kwargs):
-        assert group_type is not None
-        super().__init__(group_type, **kwargs)
+    _default_name = "Observation Points Theme"
 
     @classmethod
     def default_type_uid(cls) -> uuid.UUID:
@@ -196,13 +136,7 @@ class RockPropertiesTheme(Group):
     """The type for a INTEGRATOR Rock Properties Theme."""
 
     __TYPE_UID = uuid.UUID("{cbeb3920-a1a9-46f8-ab2b-7dfdf79c8a00}")
-
-    _name = "Rock Properties Theme"
-    _description = "Rock Properties Theme"
-
-    def __init__(self, group_type: GroupType, **kwargs):
-        assert group_type is not None
-        super().__init__(group_type, **kwargs)
+    _default_name = "Rock Properties Theme"
 
     @classmethod
     def default_type_uid(cls) -> uuid.UUID:
@@ -213,13 +147,7 @@ class SamplesTheme(Group):
     """The type for a INTEGRATOR Samples Theme."""
 
     __TYPE_UID = uuid.UUID("{1cde9996-cda7-40f0-8c20-faeb4e926748}")
-
-    _name = "Samples Theme"
-    _description = "Samples Theme"
-
-    def __init__(self, group_type: GroupType, **kwargs):
-        assert group_type is not None
-        super().__init__(group_type, **kwargs)
+    _default_name = "Samples Theme"
 
     @classmethod
     def default_type_uid(cls) -> uuid.UUID:
